+++ R/arena.R
@@ -13,6 +13,9 @@
 #' @param funnel_nbins Number of partitions for numeric columns for funnel plot
 #' @param funnel_cutoff Threshold for categorical data. Entries less frequent than specified value will be merged into one category in funnel plot.
 #' @param funnel_factor_threshold Numeric columns with lower number of unique values than value of this parameter will be treated as factors in funnel plot.
+#' @param fairness_cutoffs vector of available cutoff levels for fairness panel
+#' @param max_points_number maximum size of sample to plot scatter plots in variable against another panel
+#' @param distribution_bins vector of available bins count for histogram
 #' @param cl Cluster used to run parallel computations (Do not work in live Arena)
 #' @return Empty \code{arena_static} or \code{arena_live} class object.\cr
 #' \code{arena_static}:
@@ -55,6 +58,9 @@ create_arena <- function(live = FALSE,
                       funnel_nbins = 5,
                       funnel_cutoff = 0.01,
                       funnel_factor_threshold = 7,
+                      fairness_cutoffs = seq(0.05, 0.95, 0.05),
+                      max_points_number = 150,
+                      distribution_bins = seq(5, 40, 5),
                       cl = NULL) {
   if (live) return(
     structure(
@@ -62,32 +68,7 @@ create_arena <- function(live = FALSE,
         explainers = list(),
         # batch = one data frame of observations
         observations_batches = list(),
-        params = list(
-          cp_grid_points = grid_points,
-          ad_grid_points = grid_points,
-          ad_N = N,
-          pd_grid_points = grid_points,
-          pd_N = N,
-          fi_B = fi_B,
-          fi_n_samples = fi_N,
-          shap_B = shap_B,
-          roc_grid_points = grid_points,
-          rec_grid_points = grid_points,
-          fm_nbins = funnel_nbins,
-          fm_cutoff = funnel_cutoff,
-          fm_factor_threshold = funnel_factor_threshold
-        ),
-        timestamp = as.numeric(Sys.time())
-      ),
-      class = "arena_live"
-    )
-  )
-  else return(
-    structure(
-      list(
-        explainers = list(),
-        # batch = one data frame of observations
-        observations_batches = list(),
+        datasets = list(),
         params = list(
           cp_grid_points = grid_points,
           ad_grid_points = grid_points,
@@ -102,6 +83,39 @@ create_arena <- function(live = FALSE,
           fm_nbins = funnel_nbins,
           fm_cutoff = funnel_cutoff,
           fm_factor_threshold = funnel_factor_threshold,
+          fairness_cutoffs = fairness_cutoffs,
+          vaa_points_number = max_points_number,
+          vd_bins = distribution_bins
+        ),
+        timestamp = as.numeric(Sys.time())
+      ),
+      class = "arena_live"
+    )
+  )
+  else return(
+    structure(
+      list(
+        explainers = list(),
+        # batch = one data frame of observations
+        observations_batches = list(),
+        datasets = list(),
+        params = list(
+          cp_grid_points = grid_points,
+          ad_grid_points = grid_points,
+          ad_N = N,
+          pd_grid_points = grid_points,
+          pd_N = N,
+          fi_B = fi_B,
+          fi_n_samples = fi_N,
+          shap_B = shap_B,
+          roc_grid_points = grid_points,
+          rec_grid_points = grid_points,
+          fm_nbins = funnel_nbins,
+          fm_cutoff = funnel_cutoff,
+          fm_factor_threshold = funnel_factor_threshold,
+          fairness_cutoffs = fairness_cutoffs,
+          vaa_points_number = max_points_number,
+          vd_bins = distribution_bins,
           cl = cl
         ),
         plots_data = list()
@@ -109,68 +123,6 @@ create_arena <- function(live = FALSE,
       class = "arena_static"
     )
   )
-}
-
-#' Prints static arena summary
-#'
-#' @param x \code{arena_static} object
-#' @param ... other parameters
-#' @export
-#' @return None
-#' @examples
-#' library("DALEX")
-#' library("arenar")
-#' library("dplyr", quietly=TRUE, warn.conflicts = FALSE)
-#' # create a model
-#' model <- glm(m2.price ~ ., data=apartments)
-#' # create a DALEX explainer
-#' explainer <- DALEX::explain(model, data=apartments, y=apartments$m2.price)
-#' # prepare observations to be explained
-#' observations <- apartments[1:3, ]
-#' # rownames are used as labels for each observation
-#' rownames(observations) <- paste0(observations$construction.year, "-", observations$surface, "m2")
-#' # generate static arena for one model and 3 observations
-#' arena <- create_arena(live=FALSE) %>% push_model(explainer) %>% push_observations(observations)
-#' # print summary
-#' print(arena)
-print.arena_static <- function(x, ...) {
-  cat("===== Static Arena Summary =====\n")
-  jstr <- get_json_structure(x)
-  cat(paste("Models:", paste(jstr$models, collapse = ", "), "\n"))
-  cat(paste("Observations:", paste(jstr$observations, collapse = ", "), "\n"))
-  cat(paste("Variables:", paste(jstr$variables, collapse = ", "), "\n"))
-  cat(paste("Plots count:", length(jstr$data), "\n"))
-}
-
-#' Prints live arena summary
-#'
-#' @param x \code{arena_live} object
-#' @param ... other parameters
-#' @export
-#' @return None
-#' @examples
-#' library("DALEX")
-#' library("arenar")
-#' library("dplyr", quietly=TRUE, warn.conflicts = FALSE)
-#' # create a model
-#' model <- glm(m2.price ~ ., data=apartments)
-#' # create a DALEX explainer
-#' explainer <- DALEX::explain(model, data=apartments, y=apartments$m2.price)
-#' # prepare observations to be explained
-#' observations <- apartments[1:30, ]
-#' # rownames are used as labels for each observation
-#' rownames(observations) <- paste0(observations$construction.year, "-", observations$surface, "m2")
-#' # generate live arena for one model and 30 observations
-#' arena <- create_arena(live=TRUE) %>% push_model(explainer) %>% push_observations(observations)
-#' # print summary
-#' print(arena)
-print.arena_live <- function(x, ...) {
-  cat("===== Live Arena Summary =====\n")
-  jstr <- get_json_structure(x)
-  cat(paste("Models:", paste(jstr$models, collapse = ", "), "\n"))
-  cat(paste("Observations:", paste(jstr$observations, collapse = ", "), "\n"))
-  cat(paste("Variables:", paste(jstr$variables, collapse = ", "), "\n"))
-  cat("Remember to start server with arena_run(arena)\n")
 }
 
 #' Adds model to arena
@@ -317,6 +269,77 @@ push_observations.arena_live <- function(arena, observations) {
   arena
 }
 
+#' Adds new datasets to Arena
+#' 
+#' Adds data frame to create exploratory data analysis plots
+#' @param arena live or static arena object
+#' @param dataset data frame used for EDA plots
+#' @param target name of target variable
+#' @param label label of dataset
+#' @return Updated arena object
+#' @export
+#' @examples
+#' library("DALEX")
+#' library("arenar")
+#' library("dplyr", quietly=TRUE, warn.conflicts = FALSE)
+#' # create live arena with only one dataset
+#' apartments <- DALEX::apartments
+#' arena <- create_arena(live=TRUE) %>% push_dataset(apartments, "m2.price", "apartment")
+#' print(arena)
+#' # add another dataset
+#' HR <- DALEX::HR
+#' arena <- arena %>% push_dataset(HR, "status", "HR")
+#' print(arena)
+push_dataset <- function(arena, dataset, target, label) {
+  UseMethod("push_dataset")
+}
+
+#' @export
+#' @importFrom methods is
+push_dataset.arena_live <- function(arena, dataset, target, label) {
+  if (is.null(arena) || !is(arena, "arena_live")) {
+    stop("Invalid arena argument")
+  }
+  validate_new_dataset(arena, dataset, target, label)
+
+  # save dataset
+  n <- length(arena$datasets) + 1
+  arena$datasets[[n]] <- list(
+    dataset = dataset,
+    target = target,
+    label = label,
+    variables = colnames(dataset)[colnames(dataset) != target]
+  )
+  # update timestamp
+  arena$timestamp <- as.numeric(Sys.time())
+  arena
+}
+
+#' @export
+#' @importFrom methods is
+push_dataset.arena_static <- function(arena, dataset, target, label) {
+  if (is.null(arena) || !is(arena, "arena_static")) {
+    stop("Invalid arena argument")
+  }
+  validate_new_dataset(arena, dataset, target, label)
+
+  # save dataset
+  n <- length(arena$datasets) + 1
+  arena$datasets[[n]] <- list(
+    dataset = dataset,
+    target = target,
+    label = label,
+    variables = colnames(dataset)[colnames(dataset) != target]
+  )
+
+  # calculate eda plots and append to list
+  arena$plots_data <- c(
+    arena$plots_data,
+    get_dataset_plots(arena$datasets[[n]], arena$params)
+  )
+  arena
+}
+
 #' Upload generated json file from static arena
 #'
 #' By default function opens browser with new arena session. Appending data to
@@ -358,5 +381,27 @@ upload_arena <- function (arena, open_browser = TRUE, append_data = FALSE,
     # open new session
     utils::browseURL(paste0(arena_url, "?data=", url))
   }
+  arena
+}
+
+#' Save generated json file from static arena
+#'
+#' @param arena Static arena object
+#' @param filename Name of output file
+#' @param pretty whether to generate pretty and easier to debug JSON
+#' @return not modified arena object
+#' @export
+#' @importFrom methods is
+save_arena <- function (arena, filename="data.json", pretty=FALSE) {
+  if (is.null(arena) || !is(arena, "arena_static")) {
+    stop("Invalid arena argument")
+  }
+  # generate json string
+  json <- jsonlite::toJSON(
+    get_json_structure(arena),
+    auto_unbox = TRUE,
+    pretty = pretty
+  )
+  write(json, file = filename)
   arena
 }