+++ tests/testthat/test_plots_count.R
@@ -16,7 +16,7 @@ arena <- push_observations(arena, new_observation)
 
 n_vars <- ncol(new_observation) - 1
 n_obs <- nrow(new_observation)
-correct_len <- 4 + # FI, REC, Metrics, FM
+correct_len <- 5 + # FI, REC, Metrics, FM, SP
   n_obs * 2 + # BD, SHAP
   n_vars * 2 + # ADP, PDP
   n_vars * n_obs # CP
@@ -32,7 +32,16 @@ arena_cluster <- push_model(arena_cluster, explainer_rf)
 arena_cluster <- push_observations(arena_cluster, new_observation)
 
 test_that("regression plots count is correct using cluster", {
-  expect_equal(length(arena_cluster$plots_data), correct_len)
+  tryCatch(
+    expr = {
+      expect_equal(length(arena_cluster$plots_data), correct_len)
+    }, 
+    error = function(e){
+      print(e)
+      # if failed print table of plot types
+      print(table(sapply(arena$plots_data, function(p) p$plotType)))
+    }
+  )
 })
 
 stopCluster(cl)
@@ -50,11 +59,20 @@ arena <- push_observations(arena, new_observation)
 
 n_vars <- ncol(new_observation) - 1
 n_obs <- nrow(new_observation)
-correct_len <- 5 + # FI, REC, ROC, Metrics, FM
+correct_len <- 6 + # FI, REC, ROC, Metrics, FM, SP
   n_obs * 2 + # BD, SHAP
-  n_vars * 2 + # ADP, PDP
+  n_vars * 3 + # ADP, PDP, Fairness
   n_vars * n_obs # CP
 
 test_that("Classification plots count is correct", {
-  expect_equal(length(arena$plots_data), correct_len)
+  tryCatch(
+    expr = {
+      expect_equal(length(arena$plots_data), correct_len)
+    }, 
+    error = function(e){
+      print(e)
+      # if failed print table of plot types
+      print(table(sapply(arena$plots_data, function(p) p$plotType)))
+    }
+  )
 })